Handler of segmentation fault improvements... there is a problem yet that no free() can be called in the handler and therefore the PageResponse should be an static allocated member, it was tested this way and it worked, but it was not changed in this commit because of the way it is used in the sockets...
Another thing is still pending, regarding how to know if the fault was caused by a read or a write operation. In this moments if the faulting page was not present it is asked with read only access, if it should have been asked with write access then another fault is generated in the same instruction and the access is updated. It works but it is not ideal.